requirements.txt
azure-functions
Flask
pandas
fredapi
pyodbc
requests
matplotlib
numpy
plotly
seaborn
quandl
sklearn